backend/notification/pyproject.toml
[tool.poetry]
name = "notification"
version = "0.1.0"
description = ""
authors = ["ligets <legen2a208@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic-settings = "^2.5.2"
python-dotenv = "^1.0.1"
datetime = "^5.5"
aio-pika = "^9.4.3"
fastapi-mail = "^1.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
